SEN0377 integration and BUGFIX (#27)
* [SEN0377 Update]

First implementation of the SEN0377 drivers.
Added id into the config file.
Integrated the drivers from library

* [FW UPDATE]

Implementation of the newest SEN0377 sensor.
Upgrade and review of the FOTA procedures.
UNDER TEST

* [MICS SEN UPDATE]

Update the MICS sensor init.
Not it is detected.
Update the FW update procedure to be checked

* [CODE UPDATE]

While the connection is established only before a data transmission, the FOTA was having problems.
Now the process is working, keeping a flag active to check if new FW version are downloadeable or not.

New MICS sensor reacheable, but faulty data reading.

Tested and working

* [DOCUMENTATION UPDATE]

Improved the documentation describing how FOTA process is done