[UPDATE]
MICS sensor
MICS sensor implementation 100% done. R0 data are taken from the config file as it was in the previous MICS sensor.
Data accumulated before applying the mean and then the power-approximant formula.

Bugfix
Solved a bug in the data transmission. While there is 00:00:00 the data has previously sent, and then the device will check for firmware updates.
While starting at a time far from the standard starting-time, it calibrates the mean number to not lose any data point.

Makefile
Added a new partition table, the firmware size now is bigger than before, and with the OTA we need a custom partition table.

Extra
Added some generic functions and solved some typos and errors.

KNOWN ISSUES
while using the GSM connection the device is not currently able to manage HTTP or HTTPS connections.
So the GSM connection is NOT working