[BUGIX]
Added a debug flag to check time-related procedures, writing a fake NTP time and sync flag.
With this debug mode the device also simulates a standard transmission.

Update the data sending procedure.
Update the firmware version check procedure.
Now the device sends first the whole measurement data, then it checks for a new firmware version and, if not present also for the new NTP sync.
If a new firmware version is online, it flashes and makes the NTP update at the standard boot-time.