[MAKEFILE UPDATE]
Changed the partition tables.
Changed the makefile, now there is a FLASH_SIZE flag which, if set to 4MB or 4M will create the firmware for the 4MB flash-sized device. Otherwise it will use 8MB as default.

The makefile, while writing make env, will copy the partitions_Xmb.csv into the proper folder.
Then it will create a backup of the boards.txt file, then it will add the lines to properly detect the custom partition tables.